fix Subprocess finding libc on apple platforms
Summary: Long ago, the libc dylib used to be a file `/usr/lib/libc.dylib`. It is now not even a file - it is magic performed by the dynamic loader. But libc is referenced as `/usr/lib/libSystem.B.dylib` - this can be observed by running `otool -L ${bin}` on a binary that links libc.

Reviewed By: mattjgalloway

Differential Revision: D71628266

fbshipit-source-id: c17164c732c1e98047d5f675adb0e439041488f3