Explicitly disallow coroutines be scheduled on weak executors
Summary: Work is usually scheduled on a weak executor via a "weak" KeepAlive token that keeps a WeakRef executor alive instead of the actual thread-backed executor. When coroutines are scheduled on a weak executor, it leads to memory leak, because of a violation of contract. Coroutine frames are managed by the coroutine machinery via RAII types like the return_type, awaitor type. But there are certain cases where the RAII type explicitly gives up the ownership of the coroutine handle. It can happen when a coroutine is scheduled on an executor. An awaitor gives up the ownership of a coroutine frame and expects the executor to run the coroutine to completion and free the resources. A weak executor does nothing on `add()`, which explicitly violates the contract, and can leave the coroutine frame leaked.

Reviewed By: iahs

Differential Revision: D71625849

fbshipit-source-id: cec111b863a189ee66162d8473b17e835ece699a