Move OperationCancelled to own header, add "do not use" comment
Summary:
This achieves two goals:
 - Don't require upcoming `folly::result` to depend on the rather-larger, and entirely orthogonal `CancellationToken.h`.
 - Add a comment nudging towards non-usage of the "cancellation as an exception" paradigm. It will take a lot more than this to eliminate it for C++26 compat, but it's a baby step in that direction.

While I'm in here, `coro/Result.h` doesn't need to depend on `CancellationToken.h`. There are a couple of other unused-looking includes of `CancellationToken.h` in `folly`, but I decided not to couple this change to chasing down whomever happens to implicitly depend on them.

Reviewed By: yfeldblum

Differential Revision: D72181807

fbshipit-source-id: 802e0c40df254c60bc3867d8cbb5104a3eae38e4