Drop support for <experimental/memory_resource>
Summary:
Folly requires a c++17 compiler, but the standard library may be behind and not have `<memory_resource>`. `MemoryResource.h` falls back to `<experimental/memory_resource>` when available, but in order to support that everything else needs to use the awkward aliases in `folly::detail::std_pmr::` (even for the public interface).
This also makes it impossible to reliably forward-declare `memory_resource` (technically UB, but sometimes necessary).

This diff removes the fallback: either `<memory_resource>` is available, or we just do not provide the functionality.

Reviewed By: Gownta

Differential Revision: D72082237

fbshipit-source-id: cb1d435243e3408ec1c307fb6676040e1e6b187c