Make exception_tracer work with libc++abi (#2418)
Summary:
Pull Request resolved: https://github.com/facebook/folly/pull/2418

The exception tracer is a facility we use to display backtraces when the program (or thread) dies with an exception. It also stores the stack trace during unwinding of the exception for future retrieval should the user desire. Originally, the full exception tracer was hooked  to work with only libstdc++'s ABI implementation called `libsupc++`. However, with a few adjustments, this can be made to work with `libc++abi` and potentially libcxxrt on FreeBSD too.

There are two main changes to make exception tracer work with libc++ and libc++abi.

1. We need to dlsym the mangled name for `std::rethrow_exception(std::exception_ptr)`. There's a difference between the two libraries. We rely on the std library macro (i.e. `__GLIBCXX__` or `_LIBCPP_VERSION`) to select which one.
2. Patched `__cxa_exception` struct to include the fields to reflect the difference between the two libraries.

Reviewed By: yfeldblum

Differential Revision: D73237091

fbshipit-source-id: c7b6cff35545e81627b67d10830a8ef7528223d9