fix AsyncIoUringSocketTest under libc++
Summary:
The test is failing under libc++ because of implementation difference for `std::unique_ptr` destructor. In the libstdc++ version, in the destructor the object is destroyed first, then the pointer is changed to null. But in the libc++ version, the pointer to the object is copied to temp variable and replaced with null first, then the object is destroyed via the temp variable.

In the test case, the `unique` ptr of `AsyncSocketTransport` is passed to `EchoTransport`. `EchoTransport` registers itself with `AsyncSocketTransport` as a callback by passing a raw pointer to itself. `EchoTransport` destructor calls `unique_ptr` destructor, which destroys `AsyncSocketTransport`. `AsyncSocketTransport` destructor calls `readEOF` on the read callback (which is `EchoTransport`), `readEOF` calls `close` via the `unique_ptr` holding `AsyncSocketTransport`. However, the `unique_ptr` is already being destroyed and in libc++ the pointer to the object is already gone.

Reviewed By: grigorypas

Differential Revision: D73732983

fbshipit-source-id: a4c667805600e74b687381d667bd321dac4a27f5