AsyncSocket: add async recvzc reads
Summary:
Add async recvzc reads to AsyncSocket.

# Context

There are two socket peeks prior to an established Thrift connection. One to peek 9 bytes for TLS and another to peek 13 bytes for Thrift frame type. This is done early on in the connection establishment process, before an AsyncIoUringSocket is created. The peeks are done through Transport/SocketPeeker and use the standard libevent EventHandlers, i.e. `AsyncSocket::handleRead()`.

However with zero copy receive the only way to read data out of the socket is to use an io_uring recvzc request.

D69964064 added a way to do an io_uring recvzc request from AsyncSocket, via a new `ReadMode::ReadZC` that then calls `IoUringBackend::issueRecvZc()`. But it was buggy because it tried to make `IoUringBackend::issueRecvZc()` synchronous as a direct drop in replacement for `recv()` or `recvmsg()` used in AsyncSocket.

# Solution

## Issue Side

io_uring fundamentally cannot do synchronous requests. To get around this, add a new async read path to AsyncSocket.

In `AsyncSocket::performReadMsg()` if ReadMode is ReadZC then queue up a recvzc request, as before. But instead of expecting this to synchronously produce a read + a result, make it async by passing in a recvzc completion callback that will be invoked by IoUringBackend once the recvzc request is done.

Return a new `ReadResultEnum::READ_ASYNC` to the upper layer.

The upper layer `AsyncSocket::processNormalRead()` then returns `ReadCode::READ_DONE`.

This effectively ends `AsyncSocket::handleRead()` and `AsyncSocket::ioReady()` without calling the ReadCallback. This frees up the EVB thread now to do other work.

A DestructorGuard for the AsyncSocket must be taken to ensure it is not destroyed prior to the recvzc completion callback running.

## Completion Side

The queued up recvzc request will eventually complete, running its completion callback which then either calls `ReadCallback::readDataAvailable()` or `readErr()` depending on the completion result.

This completes the read and causes Transport/SocketPeeker to call `peekSuccess()` or `peekError()`.

Reviewed By: yfeldblum

Differential Revision: D73463675

fbshipit-source-id: 71f0d10ded61888c45c647cbb83b1f38dcb732eb