BufferedMultiSlidingWindow
Summary:
fb303 `BasicQuantileStat`, used by all quantile counters, keeps one `BufferedStat` for all-time statistics, and one for each sliding window, and each `addValue()` needs to update all of them. The vast majority of the cost comes from appending the value to the per-CPU buffers, and from merging them when full, and this work is duplicated for each `BufferedStat`. Similarly, almost all the memory comes from per-CPU buffers, which are also duplicated.

Since the all-time `BufferedStat` uses a 1-second merging schedule, and all the sliding window buckets use multiples of 1 second, we can instead use a single 1-second `BufferedStat`, and merge its digests into both all-time and windowed digests.

For all-time and minute counters, the behavior is identical since the buffer is already 1 second for both of them. For larger windows, if the digest merge operation is associative, the result is identical. `TDigest` is "associative enough", since we rely on that to distribute collection across cores anyway.

In the common case where only a 60-second window is maintained, this should almost halve both CPU and memory cost. If more windows are used, the savings are even larger.

This diff only introduces `BufferedMultiSlidingWindow`, integration with fb303 is done in the next diff.

Reviewed By: Gownta

Differential Revision: D74163724

fbshipit-source-id: 1459dc1515786784096b4a71c978a55b34898fed