Eliminate the inline assembly in folly::Instructions
Summary:
These were previously inline assembly to get GCC & Clang to inline them even when we weren't compiling with the required instruction set enabled, however these are now all included in the baseline instruction set we assume is present, so the intrinsics all get inlined properly.
Switching over to the intrinsics also has the added benefit of allowing compile time folding that isn't possible with the inline assembly. This is particularly noticable with the bextr instruction, as the inline assembly version was actually generating worse code :(

Reviewed By: Gownta

Differential Revision: D74504609

fbshipit-source-id: d1a9ceac8522bd68213c9ff3cdfdc8cfa2ff1537