Add versions of fnv32 and fnv64 that handle sign extension correctly
Summary:
The `a` variants already handle this correctly. These are not supposed to be doing sign-extension according to the ieee spec for this algorithm. This means callers of these hash functions are getting the wrong results.
This is the first of a series of diffs that will allow the default implementations of these functions to behave correctly.

Initially the correct version of these functions have the `_FIXED` suffix, but that suffix will be removed once the current, broken, implementation is renamed to have the `_BROKEN` suffix.

Reviewed By: yfeldblum

Differential Revision: D72750895

fbshipit-source-id: 8a11c9b9926fe9bf7c2aa6d20b5af72525a32188