Declare weak `malloc_stats_print` symbol
Summary:
We have declarations for malloc related symbols in the `MallocImpl.h` file, but
there is no `malloc_stats_print` there. Because of that declarations are
scattered across code base in users code.

Lets declare weak `malloc_stats_print` in one place and reuse it, same as we do
with `mallctl` and friends.

Reviewed By: ot

Differential Revision: D75525345

fbshipit-source-id: ad285ff1ce652cde75376b231f7b0103653a2062