folly::down_heap()
Summary:
When using a heap to merge sorted sets, the top needs to be updated after each entry is consumed, and the STL heap manipulation functions do not offer an operation to do that: it's necessary to pop the top, modify it, and then push it again. This is wasteful, especially in the situation where the modified top goes back to the top, which is not uncommon when merging sets with skewed sizes.

This diff adds the complementary operation `down_heap()`, which can be used together with the STL heap operations.

Reviewed By: skrueger, DenisYaroshevskiy

Differential Revision: D75446632

fbshipit-source-id: 6318ce32be3baf5dd042c6cce134e5e200194913