Remove OpaqueTaskWrapperCrtp
Summary:
This was only used in `SafeTask.h` to implement `ClosureTask<>`, but the extra abstraction wasn't carrying its own weight.

Worse, I would have to make it even more complicated in order for `AsNoexcept<ClosureTask<>> foo()` to compile. In short:
  - `AsNoexcept` needs to know `semi_await_result_t` so that it can properly wrap the promise of the underlying task
  - `semi_await_result_t` is implemented via `co_viaIfAsync`
  - specializing `semi_await_result_t` would (slightly) hurt compile times for everything, so that's not a good option
  - specializing `AsNoexcept<ClosureTask<>>` is possible, but a bunch of extra boilerplate.

The updated docblock discusses why users should't instantiate `ClosureTask`s except via async closures, but ... in the event they do, nothing terrible will happen. The main risk is that users accidentally move a `capture<Value>` into such a manually-managed `ClosureTask`, but (a) that's not actually dire, just a low-consequence violation of the usage model, (b) but this should have a linter anyway.

Reviewed By: ispeters

Differential Revision: D75566760

fbshipit-source-id: be77673f75fa9559187cb783f5f4e7a4b00142e3