`lang/named/Bindings.h`: implement `"x"_id = ...` kwargs
Summary:
This comes on the heels of `lang/Bindings.h` to show to extend bindings with new functionality.

For now, this is used for "capture backrefs" (up-stack).

This "named binding" implementation can **later** be used to implement real lexical scopes in `folly/lang/named/Scopes.h`, and `async_kwarg_closure`.

Reviewed By: ispeters

Differential Revision: D63299799

fbshipit-source-id: 237d36baded16732b887a44df0325d37e69c7b7d