Move futures.h, executor.h, coro.h to be WeakPython based
Summary:
Cython `api.h` is confusing and it doesn't work with c++ modules.  So instead we can use function ptrs and weak python to handle imports that way we no longer have global scope functions and buck configs are much easier to reason about.

Simplified the buck configs for xplat, they didn't even have an `:executor_lib` over there

I added an `executor_details.pxd` so I can control naming with import as to not conflict with existing functions in the cython code.

https://fb.workplace.com/groups/fbpython/posts/29701170732830911/

Reviewed By: itamaro

Differential Revision: D75321950

fbshipit-source-id: 16f729f08f54857d84dbcc9710ace5569486adc7