Use folly::hardware_concurrency in place of std::thread::hardware_concurrency
Summary:
This change replaces uses of std::thread::hardware_concurrency with
folly::hardware_concurrency.

The folly method returns the number of processors available to a
process and the std method returns the number of processors available
to the operating system.

In a container environment that restricts processes to a limited set
of processors, the folly method will return an accurate value for the
available concurrency while the std method will over represent it.

Reviewed By: Orvid

Differential Revision: D75723630

fbshipit-source-id: 55a18de657683bd4bb203f2e4ad9c6467382134f