AsyncFdSocket: make addSeqNum a free function
Summary:
**Context for this stack:**

Using `AsyncFdSocket` to connect to a local thrift server that uses service framework over UDS and passing header fds currently does not work. This is because the server logic depends on `PeekingAcceptorHandshakeHelper` that peeks into the socket data and if it doesn't look like TLS data forwards to `UnencryptedAcceptorHandshakeHelper`. `UnencryptedAcceptorHandshakeHelper` does not use `AsyncFdSocket` so it misses any cmsg fds. Changing `UnencryptedAcceptorHandshakeHelper` to use `AsyncFdSocket` alone is not sufficient as the initial peeking operates on `AsyncSSLSocket` and can also miss cmsg fds.

To fix this, this stack moves the queue of cmsg fds from `AsyncFdSocket` to the base `AsyncSocket` class, and sets `readAncillaryDataCallback` in `AsyncSSLSocket` to allow receiving cmsg fds when the `AsyncSSLSocket` is in the unencrypted state.

**This diff:**
Turn `AsyncFdSocket::addSeqNum` into a free function in `folly::detail` as suggested in https://www.internalfb.com/diff/D75605231?transaction_fbid=1237610058010253

Reviewed By: snarkmaster

Differential Revision: D75676137

fbshipit-source-id: 8d8e057597490092419904d12dbd95a4f6844354