refactor the folly::coro::SharedMutex to have self-contained state transition functions
Summary:
This diff refactors the code to make it more clear where the state transition happens for the SharedMutex state machine. This is in preparation for the support to make SharedMutex upgradeable, which will introduce a few more states.

This `unlockOrGetNextWaitersToResume` function used to be only called when the state is effectively unlocked and it looks for the next waiters to unblock. If there are none, it actualizes the unlock state via assignment. The distance between action (e.g. unlock()) and state transition (e.g. setting state to unlocked) is far and split in two different functions.

Now we make each state transition function self-contained. `unlock_shared`/`unlock` both perform the complete state transition including state assignment before calling `getWaitersToResume` which may perform another state transition. If the new state is not unlocked, nothing happens. Otherwise, it looks for the next waiters to unblock in the same way.

`getWaitersToResume` is performed with the same lock held from `unlock_shared`/`unlock`. So even though logically we perform two state transitions, they are done atomically.

Reviewed By: Gownta, yfeldblum

Differential Revision: D75332419

fbshipit-source-id: 3d0da546be5dbc7ed9a898da9a34a8dc9d2d5ffe