add try_lock_upgrade() and unlock_upgrade()
Summary: Add `try_lock_upgrade()` and `unlock_upgrade()` methods, and introduced a new mutex state field `waitingWriterCount_` which keeps track of the number of waiting writers. Before coro::SharedMutex supports upgrade lock, `lock_shared()` will wait until the mutex is not exclusively locked _and_ there's no waiter. Before upgrade lock, the only waiter possible is the writer lock (e.g. the mutex has active shared locks, and one or more writer locks are waiting). Because the mutex is fair/writer prioritized, we do not want to grant new readers shared lock to prevent writer starvation. However, with the introduction of upgrade lock, the "no waiter" check is too strict. E.g. we want to grant new readers shared locks when the mutex is upgrade locked, and even there are other pending upgrade lock requests, because upgrade lock and shared locks do not conflict. What we really want, is not to grant any readers shared lock if there are writers waiting. This is the case before and after the introduction of the upgrade lock. In this diff, we start tracking the number of waiting writers and update the canLockShared logic accordingly.

Reviewed By: yfeldblum, Gownta

Differential Revision: D75415606

fbshipit-source-id: 69bf84a60ef25b9813b498c49f3e2ca1a12a1f21