Optimize locking behavior of accessAllThreads
Summary:
accessAllThreads needs to acquire several locks to keep the state of
the TL object being iterated stable across all threads during the lifetime of
the Accessor. The StaticMeta::lock_ prevents a thread's element array from
being reallocated due to it updating some, potentially unrelated, TL object.
The Accessor needs to read the element ptr from each thread's ThreadEntry
element array. The ThreadEntrySet for the TL object being iterated is also
locked to prevent any thread from altering the element it has for that TL
object while the Accessor is alive. By caching the element ptr in the
ThreadEntrySet, we eliminate the need for the Accessor to rely on `elements`
array in ThreadEntry and hence the need to acquire StaticMeta's lock_.

Besides simplifying the interactions between different parts of the
implementation, this change will also prevent false blocking of threads that
want to update TL object unrelated to the Accessor. It opens the path to
moving the element deleter info to ThreadEntrySet and further reduce per
element memory overheads.

Reviewed By: yfeldblum

Differential Revision: D76320828

fbshipit-source-id: 24065a1a24fd45a73248bc40a0cf5e6f8000ba49