Minor code move and add a DCHECK
Summary:
The code to allocate and configure a new threadEntry is moved to a new
allocateThreadEntry() member function. This will allow for adding better sanity
checking to ensure TL are not used in a way that isn't possible to support
well. Specifically, today there is no checks against creating new ThreadEntry
during thread exit handlers. This can result in leaks. Added a check to detect
this in debug builds so we can fix calling code behavior.

run_all_tests

Reviewed By: yfeldblum

Differential Revision: D76757950

fbshipit-source-id: a31a9f30d92d75af184ac0ac455dd37b691ee647